+++ src/testcurl/daemon_options_test.c
@@ -40,7 +40,7 @@ ahc_echo (void *cls,
           const char *url,
           const char *method,
           const char *version,
-          const char *upload_data, unsigned int *upload_data_size,
+          const char *upload_data, size_t *upload_data_size,
           void **unused)
 {
   return 0;
