+++ src/testcurl/https/tls_test_common.c
@@ -23,7 +23,6 @@
  * @brief  Common tls test functions
  * @author Sagie Amir
  */
-
 #include "tls_test_common.h"
 #include "tls_test_keys.h"
 #include "gnutls.h"
@@ -55,7 +54,7 @@ copyBuffer (void *ptr, size_t size, size_t nmemb, void *ctx)
 }
 
 static int
-file_reader (void *cls, size_t pos, char *buf, int max)
+file_reader (void *cls, uint64_t pos, char *buf, int max)
 {
   FILE *file = cls;
   fseek (file, pos, SEEK_SET);
@@ -68,7 +67,7 @@ file_reader (void *cls, size_t pos, char *buf, int max)
 int
 http_ahc (void *cls, struct MHD_Connection *connection,
           const char *url, const char *method, const char *upload_data,
-          const char *version, unsigned int *upload_data_size, void **ptr)
+          const char *version, size_t *upload_data_size, void **ptr)
 {
   static int aptr;
   struct MHD_Response *response;
@@ -112,7 +111,7 @@ http_ahc (void *cls, struct MHD_Connection *connection,
 int
 http_dummy_ahc (void *cls, struct MHD_Connection *connection,
                 const char *url, const char *method, const char *upload_data,
-                const char *version, unsigned int *upload_data_size,
+                const char *version, size_t *upload_data_size,
                 void **ptr)
 {
   return 0;
